refactor state color computation
